Fixed bad typing of strong_threshold, changing it from long int * to double *

+++ parcsr_ls/F90_HYPRE_parcsr_amg.c
@@ -122,7 +122,7 @@ hypre_F90_IFACE(hypre_paramgsetmaxlevels)( long int *solver,
 
 void
 hypre_F90_IFACE(hypre_paramgsetstrongthreshold)( long int *solver,
-                                                 long int *strong_threshold,
+                                                 double   *strong_threshold,
                                                  int      *ierr              )
 {
    *ierr = (int)
